Fixed tests from mapreduce refactor
