--- bin/src/server/script/tests/mod.rs
@@ -1,2 +1,0 @@
-mod execute;
-mod mapreduce;
